Add files via upload
This file provides harmonised data for the UK's main manufacturing industry based on the NAEI 2020 data, which can be used for analysis and modelling. Since the original NAEI data include sites that do not fall within the definition of manufacturing industry, we created this harmonised data for ease of use. We classified the sectors into Iron and steel, Chemicals, Cement, Food and drink, Paper, Refining, Lime, Glass, Ceramics, Other minerals, and Other industry. We also converted the 'Easting' and 'Northing' to 'Latitude' and 'Longitude'.

For the Iron and steel sector, we grouped the ‘Port Talbot’ and ‘Scunthorpe’ emissions in the NAEI data into one total emission for each site. The rest is left as is. 

For the original ‘Other minerals industries’ in the NAEI data, we divided this into ‘Glass’, ‘Ceramics’, and ‘Other minerals’, based on a basic search for the site operators. Thus, this may not accurately describe the real sector for those sites.